requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.0rc4
certifi==2019.3.9
chardet==3.0.4
clearbit==0.1.7
defusedxml==0.5.0
Django==2.1.7
django-rest-auth==0.9.3
django-rest-registration==0.4.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.8
kombu==4.4.0
oauthlib==3.0.1
pyhunter==0.3
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
vine==1.2.0
